+++ Firmware/rtos_pilot/ahrs.h
@@ -1,7 +1,7 @@
 #ifndef AHRS_H
 #define AHRS_H
 
-void ahrs_filter(double dt);
+void ahrs_filter(float dt);
 void ahrs_init();
